requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0
gunicorn==22.0.0
python-dotenv==1.0.1
strawberry-graphql[fastapi]==0.230.0
liboqs-python==0.14.0
bcrypt==4.2.0
python-jose[cryptography]==3.4.0
pyjwt==2.10.1
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
tensorflow==2.16.1
prophet==1.1.5
shap==0.45.0
transformers[torch]==4.38.0
qiskit==1.0.0
qiskit-aer==0.13.0
pytesseract==0.3.10
stable-baselines3==2.3.2
statsmodels==0.14.1
asyncpg==0.29.0
pymongo==4.6.0
influxdb-client==1.40.0
redis==5.0.1
boto3==1.34.0
kafka-python==2.0.2
alembic==1.13.1
pytest==8.2.2
pytest-asyncio==1.0.0
httpx==0.27.0
locust==2.24.0
ruff==0.6.0
prometheus-fastapi-instrumentator==6.1.0
fastapi-limiter==0.1.6
schedule==1.2.0
web3==6.16.0
pubchempy==1.0.4
cmdstanpy==1.2.0
structlog==24.1.0
pydantic-settings==2.3.4